requirements.txt
# Trading Bot Dashboard - Requirements for Streamlit Cloud
# ========================================================

# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0

# Technical Analysis
ta>=0.10.0

# Machine Learning (optional - for ML features)
scikit-learn>=1.3.0
xgboost>=2.0.0

# Data Loading
openpyxl>=3.1.0
requests>=2.31.0

# Logging
loguru>=0.7.0

# YAML config
pyyaml>=6.0
